dask_compat.py
For this latest version of the script, some new features and corrections of previous bugs have been added with the aim to increase the script’s usability and to simplify its code. First, the error handling was greatly improved in terms of being able to provide messages that are much more informative to the user why a certain situation is not doable and the user can easily pinpoint what went wrong. Furthermore, the same formula has been improved as before by adding conditions for invalid or missing version string that would cause the application to crash and that will make sure that the dependencies are well checked. The treatment of deprecation warnings has also been enhanced to provide a more direction on using the latest ‘importlib_metadata. entry_points’ function. In addition, improved the code readability and made consistent using of f-string for string formatting to convey the information, as well as added comments for the script through the code. Altogether these changes improve the script and prepare it better for the future updates and makes sure that they are more compatible and easy to integrate in the Dask ecosystem.